Fix activities
