+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         </activity>
         <activity android:name=".Users.Controllers.LoginActivity" />
         <activity android:name=".Controllers.MainActivity" />
-        <activity android:name=".Users.Controllers.LoginActivity$RegisterActivity" />
+        <activity android:name=".Users.Controllers.RegisterActivity" />
     </application>
 
 </manifest>
